refactor: rename map component to trailmap
