+++ components/index.tsx
@@ -9,7 +9,7 @@ export * from "./Landing";
 
 import dynamic from "next/dynamic";
 
-export const Map = dynamic(() => import("./Map"), {
+export const TrailMap = dynamic(() => import("./TrailMap"), {
   ssr: false,
 });
